Changed the layout of the application with AuiNotebook for better ui interactivity
